+++ CHANGELOG.md
@@ -69,7 +69,7 @@ See the [docs](https://github.com/mobxjs/mobx-utils#asyncaction) for details, bu
 ```javascript
 import {asyncAction} from "mobx-utils"
 
-mobx.configure({ enforceActions: true }) // don't allow state modifications outside actions
+mobx.configure({ enforceActions: "observed" }) // don't allow state modifications outside actions
 
 class Store {
 	@observable githubProjects = []